deepseek-ocr/requirements.txt
acres==0.5.0
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.0
configobj==5.0.9
configparser==7.2.0
distro==1.9.0
dotenv==0.9.9
etelemetry==0.3.1
filelock==3.20.0
fitz==0.0.1.dev2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
jiter==0.11.1
looseversion==1.3.0
lxml==6.0.2
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.4
openai==2.6.1
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
pillow==12.0.0
prov==2.1.1
puremagic==1.30
pydantic==2.12.3
pydantic-core==2.41.4
pydot==4.0.1
pymupdf==1.26.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyxnat==1.6.3
rdflib==7.4.0
requests==2.32.5
scipy==1.16.3
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
traits==7.0.2
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
